Fix type truncation warning that breaks build on Windows
Inside `UnsignedUnitTest`, we are passing `unit32_t` into the `shift` argument
of several opcodes (e.g., `OpCode::Rol`), which are `unit8_t`. Since we
interpret warnings as errors, this breaks the build on Windows.

This commit will change the argument type to be `uint8_t`, which fixes the
Windows build.