Disable flaky tests that prevent us from having MacOS CI
See https://github.com/BitFunnel/BitFunnel/issues/298 for more discusison. Mike
has a fix in mind for this, but doesn't have time to implement in the near
future, so I'm disabling these tests in order to allow us to enable MacOS CI and
generally avoid flaky test failures on Mac.